requirements.txt
fastapi
uvicorn
beanie
pydantic[email]
motor
python-dotenv

